Make deploy_docs.sh check resolvers first
Two changes here - the first is to add 'set -e' to the deploy_docs.sh
script which will abort the script if an error is encountered at any
point during execution. This should hopefully prevent any half-complete
deployments.

Second is to check that github.com resolves before doing anything. If it
can't be resolved it'll bail out and add a message to the system log.